// src/pages/Authorization/common/common.types.ts
export type BackgroundWrapperPropsType = {
    page?: string
}

export type QuestionLinkWrapperPropsType = {
    page?: string
}

export type InputErrorTextPropsType = {
    type?: string
}

export type ShowPasswordIconPropsType = {
    isShow: boolean
}

export type SignInSubmitValues = {
    userName: string
    password: string
}

export type SignUpSubmitValues = {
    firstName: string
    lastName: string
    userName: string
    password: string
    confirmPassword: string
}